requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
mysql-connector-python==9.0.0:q!

# python -m venv venv
# source venv/bin/activate
